# docs/swagger.yaml
basePath: /api
definitions:
...
  cases.ActionRequest:
    properties:
      comment:
        description: Optional comment shown in history
        maxLength: 500
        type: string
    type: object
  cases.CaseHistoryDTO:
    properties:
      action:
        type: string
      actor_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      new_status:
        $ref: '#/definitions/models.CaseStatus'
      old_status:
        $ref: '#/definitions/models.CaseStatus'
      reason:
        type: string
    type: object
  cases.CaseListItem:
    properties:
      category:
        type: string
      created_at:
        type: string
      id:
        type: string
      quotes:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
  cases.CreateCaseRequest:
    properties:
      category:
        maxLength: 40
        type: string
      description:
        maxLength: 2000
        type: string
      title:
        maxLength: 120
        minLength: 3
        type: string
    required:
    - category
    - title
    type: object
  cases.MarketCaseItem:
    properties:
      category:
        type: string
      created_at:
        type: string
      has_my_quote:
        description: FE can use this to disable "submit quote"
        type: boolean
      id:
        type: string
      preview:
        type: string
      title:
        type: string
    type: object
  cases.PageCases:
    properties:
      items:
        items:
          $ref: '#/definitions/cases.CaseListItem'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      pages:
        type: integer
      total:
        type: integer
    type: object
  cases.PageMarketCases:
    properties:
      items:
        items:
          $ref: '#/definitions/cases.MarketCaseItem'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      pages:
        type: integer
      total:
        type: integer
    type: object
  models.CaseStatus:
    enum:
    - open
    - engaged
    - closed
    - cancelled
    type: string
    x-enum-varnames:
    - CaseOpen
    - CaseEngaged
    - CaseClosed
    - CaseCancelled
  models.ErrorResponse:
    properties:
      code:
        description: machine-friendly error code
        example: FORBIDDEN
        type: string
      error:
        description: always true for error cases
        example: true
        type: boolean
      message:
        description: human-readable error message
        example: Forbidden
        type: string
    type: object
...
  models.ValidationErrorResponse:
    properties:
      errors:
        additionalProperties:
          items:
            type: string
          type: array
        description: field name -> list of error messages
        type: object
      message:
        example: Validation failed
        type: string
    type: object
...
info:
  contact:
    email: aldoetobex@gmail.com
    name: Aldo Rifki Putra
  description: 'API for a mini legal marketplace: clients post cases, lawyers submit
    quotes, clients accept & pay, and lawyers access files via signed URLs.'
  title: Mini Legal Marketplace API
  version: "1.0"
paths:
  /cases:
    post:
      consumes:
      - application/json
      description: Client creates a new case
      parameters:
      - description: Case payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cases.CreateCaseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: id
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create case
      tags:
      - cases
  /cases/{id}/cancel:
    post:
      consumes:
      - application/json
      description: Client cancels their own case (only if still open)
      parameters:
      - description: case id (uuid)
        in: path
        name: id
        required: true
        type: string
      - description: Optional comment
        in: body
        name: payload
        schema:
          $ref: '#/definitions/cases.ActionRequest'
      responses:
        "200":
          description: status
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel case
      tags:
      - cases
  /cases/{id}/close:
    post:
      consumes:
      - application/json
      description: Client closes their own case (only if engaged)
      parameters:
      - description: case id (uuid)
        in: path
        name: id
        required: true
        type: string
      - description: Optional comment
        in: body
        name: payload
        schema:
          $ref: '#/definitions/cases.ActionRequest'
      responses:
        "200":
          description: status
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Close case
      tags:
      - cases
...
  /cases/{id}/history:
    get:
      description: List case status changes (owner & accepted lawyer only)
      parameters:
      - description: case id (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/cases.CaseHistoryDTO'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Case history
      tags:
      - cases
...
  /cases/mine:
    get:
      description: Client lists their own cases (paginated)
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cases.PageCases'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my cases
      tags:
      - cases
...
  /marketplace:
    get:
      description: Lawyer browses OPEN cases (server-side filters & pagination; no
        client identity)
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      - description: category
        in: query
        name: category
        type: string
      - description: YYYY-MM-DD (Asia/Singapore)
        in: query
        name: created_since
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cases.PageMarketCases'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Marketplace (anonymized)
      tags:
      - marketplace
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: 'Format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"